perf(frontend): optimize bundle performance with code splitting, lazy loading, WebP images, and budget monitoring (#225)
