+++ frontend/app/api/user-profile/route.ts
@@ -15,7 +15,7 @@ export async function GET(req: NextRequest) {
     .maybeSingle()
 
   if (error) return NextResponse.json({ error: error.message }, { status: 500 })
-  return NextResponse.json(data)
+  return NextResponse.json(data, { headers: { "Cache-Control": "private, no-cache" } })
 }
 
 export async function POST(req: NextRequest) {
