+++ frontend/app/layout.tsx
@@ -13,6 +13,7 @@ import { Suspense } from "react"
 import { TxQueueBadge } from "@/components/tx-queue-badge"
 import { ScrollToTop } from "@/components/ui/scroll-to-top"
 import { TransactionRecoveryProvider } from "@/components/transaction-recovery-provider"
+import { WebVitals } from "@/components/web-vitals"
 
 export const metadata: Metadata = {
   title: "JointSave — Decentralized Community Savings on Stellar",
@@ -57,6 +58,7 @@ export default function RootLayout({
           </Suspense>
         </ThemeProvider>
         <Analytics />
+        <WebVitals />
         <Toaster />
         <TxQueueBadge />
       </body>
